# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClubHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/me:
    get:
      operationId: getCurrentUser
      tags: [user]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/settings:
    get:
      operationId: getUserSettings
      tags: [user]
      summary: Get user settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettingsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUserSettings
      tags: [user]
      summary: Update user settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettingsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clubs/{id}/events:
    get:
      operationId: getClubEvents
      tags: [events]
      summary: Get upcoming events for a club
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsListResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createClubEvent
      tags: [events]
      summary: Create an event for a club (leaders only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEventResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}:
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (leaders only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
      required:
        - success
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        graduation_year:
          type: integer
          nullable: true
      required:
        - id
        - email
        - full_name
    UserMeResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - success
        - user
    UserSettings:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        notifications_email:
          type: boolean
        notifications_reminders:
          type: boolean
        notifications_new_clubs:
          type: boolean
        notifications_chat:
          type: boolean
        notifications_digest:
          type: boolean
        notifications_push_mobile:
          type: boolean
      required:
        - full_name
        - email
        - notifications_email
        - notifications_reminders
        - notifications_new_clubs
        - notifications_chat
        - notifications_digest
        - notifications_push_mobile
    UserSettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        settings:
          $ref: "#/components/schemas/UserSettings"
      required:
        - success
        - settings
    UpdateUserSettingsBody:
      type: object
      properties:
        full_name:
          type: string
        notifications_email:
          type: boolean
        notifications_reminders:
          type: boolean
        notifications_new_clubs:
          type: boolean
        notifications_chat:
          type: boolean
        notifications_digest:
          type: boolean
        notifications_push_mobile:
          type: boolean
...
    EventSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        event_date:
          type: string
        event_time:
          type: string
        location:
          type: string
        description:
          type: string
      required:
        - id
        - title
        - event_date
        - event_time
        - location
        - description
...
    EventsListResponse:
      type: object
      properties:
        success:
          type: boolean
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
      required:
        - success
        - events
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        event_date:
          type: string
        event_time:
          type: string
        location:
          type: string
        description:
          type: string
      required:
        - title
        - event_date
        - event_time
        - location
    CreateEventResponse:
      type: object
      properties:
        success:
          type: boolean
        event_id:
          type: integer
      required:
        - success
        - event_id
...